+++ polypheny.py
@@ -11,6 +11,7 @@ import value_pb2
 import grpc
 import rpc
 from exceptions import *
+from serialize import *
 
 apilevel = '2.0'
 threadsafety = 0
@@ -90,7 +91,7 @@ class Connection:
 
 
     def cursor(self):
-        if self.chan is None:
+        if self.con is None:
             raise ProgrammingError('Connection is closed')
         cur = Cursor(self)
         self.cursors.add(cur)
@@ -114,10 +115,14 @@ class Connection:
         if self.con is None:
             assert len(self.cursors) == 0
             return
-        self.rollback()
         for cur in list(self.cursors):  # self.cursors is materialized because cur.close modifies it
             cur.close()
         assert len(self.cursors) == 0
+
+        try:
+            self.rollback()
+        except:
+            pass
         self.con.close()
         self.con = None
 
@@ -130,7 +135,7 @@ class ResultCursor:
         self.frame = frame
         restype = self.frame.WhichOneof('result')
         if restype is None:
-            # TODO when does this happen?
+            # TODO does this happen?
             self.closed = True
             return
         if restype == 'relational_frame':
@@ -148,10 +153,11 @@ class ResultCursor:
     def close(self):
         if self.closed:
             return
-        assert self.con.stub != None
-        r = statement_requests_pb2.CloseStatementRequest()
-        r.statement_id = self.statement_id
-        self.con.stub.CloseStatement(r, metadata=[('clientuuid', self.con.uuid)])
+        assert self.con.con != None
+        try:
+            self.con.con.close_statement(self.statement_id)
+        except Exception:
+            pass
         self.con = None
         self.closed = True
 
@@ -166,128 +172,10 @@ class ResultCursor:
             return self.nextframe()
 
     def nextframe(self):
-        r = statement_requests_pb2.FetchRequest()
-        r.statement_id = self.statement_id
-        self.frame = self.con.stub.FetchResult(r, metadata=[('clientuuid', self.con.uuid)])
+        self.frame = self.con.con.fetch(self.statement_id, None) # TODO
         self.rows = iter(self.frame.relational_frame.rows)  # TODO result must not be relational
         return next(self.rows)  # TODO: What happens if this returns StopIteration, but another frame could be fetched?
 
-
-# See ProtoValueDeserializer
-def py2proto(value, v=None):
-    if v is None:
-        v = value_pb2.ProtoValue()
-    if type(value) == bool:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.BOOLEAN
-        v.boolean.boolean = value
-    elif type(value) == int:
-        if -2 ** 31 <= value <= 2 ** 31 - 1:
-            # v.type = value_pb2.ProtoValue.ProtoValueType.INTEGER
-            v.integer.integer = value
-        elif -2 ** 63 <= value <= 2 ** 63 - 1:
-            # v.type = value_pb2.ProtoValue.ProtoValueType.LONG
-            v.long.long = value
-        else:
-            # v.type = value_pb2.ProtoValue.ProtoValueType.BIG_DECIMAL
-            n = ((value.bit_length() - 1) // 8) + 1  # TODO: Does this work for negative numbers?
-            v.big_decimal.unscaled_value = value.to_bytes(n, byteorder='big', signed=True)
-            v.big_decimal.scale = 0
-            v.big_decimal.precision = 0
-            print(v.big_decimal)
-    elif type(value) == bytes:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.BINARY
-        v.binary.binary = value
-        # TODO: Date
-    elif type(value) == float:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.DOUBLE
-        v.double.double = value
-        # TODO: Use BigDecimal as well?
-    elif type(value) == str:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.VARCHAR
-        v.string.string = value
-        # TODO: Time, Timestamp
-    elif value is None:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.NULL
-        v.null.CopyFrom(value_pb2.ProtoNull())
-    elif type(value) == list:
-        # v.type = value_pb2.ProtoValue.ProtoValueType.LIST
-        for element in value:
-            v.list.values.append(py2proto(element))
-    elif type(value) == dict:  # experiment to test the server with unset values
-        pass
-    else:
-        raise NotImplementedError
-
-    return v
-
-
-def proto2py(value):
-    name = value.WhichOneof("value")
-    assert name is not None
-    if name == "boolean":
-        return value.boolean.boolean
-    elif name == "integer":
-        return value.integer.integer
-    elif name == "long":
-        return value.long.long
-    elif name == "binary":
-        return value.binary.binary
-    elif name == "date":
-        raise NotImplementedError()
-    elif name == "double":
-        return value.double.double
-    elif name == "float":
-        return value.float.float
-    elif name == "string":
-        return value.string.string
-    elif name == "time":
-        raise NotImplementedError()
-    elif name == "null":
-        return None
-    elif name == "big_decimal":
-        print(value)
-        raw = value.big_decimal.unscaled_value
-        scale = value.big_decimal.scale
-        prec = value.big_decimal.precision
-
-        if scale > (2**31) - 1:
-            scale = scale - 2**32
-
-        i = int.from_bytes(raw, byteorder='big', signed=True)
-        print(f'i: {i}')
-        i = i * 10 ** (-scale)
-        print(f'i: {i} {2**77}')
-        return round(i, prec + 1)  # TODO: Round Up/Down?
-    elif name == "interval":
-        raise NotImplementedError()
-    elif name == "user_defined_type":
-        raise NotImplementedError()
-    elif name == "file":
-        raise NotImplementedError()
-    elif name == "list":
-        return list(map(lambda v: proto2py(v), value.list.values))
-    elif name == "map":
-        raise NotImplementedError()
-    elif name == "document":
-        res = {}
-        for entry in value.document.entries:
-            k = proto2py(entry.key)
-            assert isinstance(k, str)  # TODO: Correct?
-            v = proto2py(entry.value)
-            res[k] = v
-        return res
-    elif name == "node":
-        raise NotImplementedError()
-    elif name == "edge":
-        raise NotImplementedError()
-    elif name == "path":
-        raise NotImplementedError()
-    elif name == "graph":
-        raise NotImplementedError()
-    elif name == "row_id":
-        raise NotImplementedError()
-    else:
-        raise RuntimeError("Unhandled value type")
 
 
 class Cursor:
@@ -336,7 +224,7 @@ class Cursor:
             self.description.append(
                 (column.column_label, None, None, None, None, column.precision, column.scale, column.is_nullable))
 
-    def execute(self, query: str, params: Union[List[Any], Dict[str, Any]] = None, *, fetch_size: int = None, ddl_hack: bool = False) -> None:
+    def execute(self, query: str, params: Union[List[Any], Dict[str, Any]] = None, *, fetch_size: int = None) -> None:
         """
         Executes a SQL query.
 
@@ -351,48 +239,32 @@ class Cursor:
         self.reset()
 
         if params is None:  # Unparameterized query
-            req = statement_requests_pb2.ExecuteUnparameterizedStatementRequest()
-            req.language_name = 'sql'
-            req.statement = query
-            if fetch_size:
-                req.fetch_size = fetch_size
-            resp = self.con.stub.ExecuteUnparameterizedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            next(resp)  # Contains only statement_id
-            if ddl_hack:
-                return
             try:
-                r = next(resp)
+                r = self.con.con.execute_unparameterized_statement('sql', query, fetch_size)
                 assert r.HasField("result")  # Is this always true?
                 # self.rowcount = r.result.scalar # Can this be relied upon?
-                assert r.result.HasField("frame")  # Is this always true?
-                self.derive_description(r.result.frame.relational_frame)
-                self.result = ResultCursor(self.con, r.statement_id, r.result.frame)
+                if r.result.HasField("frame"):  # TODO Better Error when one of the fetch* methods is invoked.  Empty fake result?
+                    self.derive_description(r.result.frame.relational_frame)
+                    self.result = ResultCursor(self.con, r.statement_id, r.result.frame)
+                return
             except Exception as e:
                 print(e)
-            return
+                raise e
 
-        req = statement_requests_pb2.PrepareStatementRequest()
-        req.language_name = "sql"
-        req.statement = query
         if type(params) == list or type(params) == tuple:
-            resp = self.con.stub.PrepareIndexedStatement(req, metadata=[('clientuuid', self.con.uuid)])
+            resp = self.con.con.prepare_indexed_statement('sql', query)
             statement_id = resp.statement_id
-            req = statement_requests_pb2.ExecuteIndexedStatementRequest()
-            req.statement_id = statement_id
-            req.parameters.parameters.extend(list(map(py2proto, params)))
-            resp = self.con.stub.ExecuteIndexedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            self.derive_description(resp.frame.relational_frame)
-            self.result = ResultCursor(self.con, statement_id, resp.frame)
+            resp = self.con.con.execute_indexed_statement(statement_id, params, fetch_size)
+            if resp.HasField("frame"): # TODO same as above
+                self.derive_description(resp.frame.relational_frame)
+                self.result = ResultCursor(self.con, statement_id, resp.frame)
         elif type(params) == dict:
-            resp = self.con.stub.PrepareNamedStatement(req, metadata=[('clientuuid', self.con.uuid)])
+            resp = self.con.con.prepare_named_statement('sql', query)
             statement_id = resp.statement_id
-            req = statement_requests_pb2.ExecuteNamedStatementRequest()
-            req.statement_id = statement_id
-            for k, v in params.items():
-                py2proto(v, req.parameters.parameters[k])
-            resp = self.con.stub.ExecuteNamedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            self.derive_description(resp.frame.relational_frame)
-            self.result = ResultCursor(self.con, statement_id, resp.frame)
+            resp = self.con.con.execute_named_statement(statement_id, params, fetch_size)
+            if resp.HasField("frame"): # TODO same as above
+                self.derive_description(resp.frame.relational_frame)
+                self.result = ResultCursor(self.con, statement_id, resp.frame)
         else:
             raise Error("Unexpected type for params " + str(type(params)))
 
@@ -417,20 +289,13 @@ class Cursor:
         self.reset()
 
         if params is None:  # Unparameterized query
-            req = statement_requests_pb2.ExecuteUnparameterizedStatementRequest()
-            req.language_name = lang
-            req.statement = query
-            if fetch_size:
-                req.fetch_size = fetch_size
-            resp = self.con.stub.ExecuteUnparameterizedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            next(resp)  # Contains only statement_id
             try:
-                r = next(resp)
+                r = self.con.con.execute_unparameterized_statement(lang, query, fetch_size)
                 assert r.HasField("result")  # Is this always true?
                 # self.rowcount = r.result.scalar # Can this be relied upon?
-                assert r.result.HasField("frame")  # Is this always true?
-                self.derive_description(r.result.frame.relational_frame)
-                self.result = ResultCursor(self.con, r.statement_id, r.result.frame)
+                if r.result.HasField("frame"):  # TODO Better Error when one of the fetch* methods is invoked.  Empty fake result?
+                    self.derive_description(r.result.frame.relational_frame)
+                    self.result = ResultCursor(self.con, r.statement_id, r.result.frame)
             except Exception as e:
                 print(e)
             return
@@ -445,8 +310,9 @@ class Cursor:
             req.statement_id = statement_id
             req.parameters.parameters.extend(list(map(py2proto, params)))
             resp = self.con.stub.ExecuteIndexedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            self.derive_description(resp.frame.relational_frame)
-            self.result = ResultCursor(self.con, statement_id, resp.frame)
+            if r.result.HasField("frame"): # TODO same as above
+                self.derive_description(resp.frame.relational_frame)
+                self.result = ResultCursor(self.con, statement_id, resp.frame)
         elif type(params) == dict:
             resp = self.con.stub.PrepareNamedStatement(req, metadata=[('clientuuid', self.con.uuid)])
             statement_id = resp.statement_id
@@ -455,8 +321,9 @@ class Cursor:
             for k, v in params.items():
                 py2proto(v, req.parameters.parameters[k])
             resp = self.con.stub.ExecuteNamedStatement(req, metadata=[('clientuuid', self.con.uuid)])
-            self.derive_description(resp.frame.relational_frame)
-            self.result = ResultCursor(self.con, statement_id, resp.frame)
+            if r.result.HasField("frame"): # TODO same as above
+                self.derive_description(resp.frame.relational_frame)
+                self.result = ResultCursor(self.con, statement_id, resp.frame)
         else:
             raise Error("Unexpected type for params " + str(type(params)))
