+++ test_serialize.py
@@ -5,68 +5,68 @@ import pytest
 from test_helper import con, cur
 
 def test_serialize_bool(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BOOLEAN, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (True,), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BOOLEAN, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (True,))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == True
     assert cur.fetchone() is None
 
 def test_serialize_number(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BIGINT NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BIGINT NOT NULL, PRIMARY KEY(i))')
     ints = {1, 2**42}
     for i in ints:
-        cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (i,), ddl_hack=True)
+        cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (i,))
     cur.execute('SELECT a FROM t')
     res = set(map(lambda x: x[0], cur.fetchall()))
     assert ints == res
 
 def test_serialize_decimal(cur):
     pytest.skip('Broken')
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a DECIMAL NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (2**77,), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a DECIMAL NOT NULL, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (2**77,))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == 2**77
     assert cur.fetchone() is None
 
 def test_serialize_string(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a VARCHAR(255) NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', ('Hello World',), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a VARCHAR(255) NOT NULL, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', ('Hello World',))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == 'Hello World'
     assert cur.fetchone() is None
 
 def test_serialize_binary(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a FILE NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (b'Hello World',), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a FILE NOT NULL, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (b'Hello World',))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == b'Hello World'
     assert cur.fetchone() is None
 
 def test_serialize_float(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a DOUBLE NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (2.71,), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a DOUBLE NOT NULL, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (2.71,))
     cur.execute('SELECT a FROM t')
     assert math.isclose(cur.fetchone()[0], 2.71)
     assert cur.fetchone() is None
 
 def test_serialize_null(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a INTEGER, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (None,), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a INTEGER, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (None,))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == None
     assert cur.fetchone() is None
 
 def test_serialize_list(cur):
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a ARRAY, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', ([1, 2, 3],), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a ARRAY, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', ([1, 2, 3],))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == [1, 2, 3]
     assert cur.fetchone() is None
@@ -110,23 +110,32 @@ def test_fail_with_superfluous_param(cur):
 
 def test_no_error_when_invalid_create(cur):
     pytest.skip("Does not properly error out")
-    cur.execute('CREATE TABLE t(a BOOLEAN)', ddl_hack=True)
+    cur.execute('CREATE TABLE t(a BOOLEAN)')
 
 def test_serialize_null_string(cur):
     pytest.skip('Broken')
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a VARCHAR(255), PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (None,), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a VARCHAR(255), PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (None,))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == None
     assert cur.fetchone() is None
 
 def test_serialize_varbinary(cur):
     pytest.skip('Causes a HSQLDB exception')
-    cur.execute('DROP TABLE IF EXISTS t', ddl_hack=True)
-    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BINARY VARYING NOT NULL, PRIMARY KEY(i))', ddl_hack=True)
-    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (b'Hello World',), ddl_hack=True)
+    cur.execute('DROP TABLE IF EXISTS t')
+    cur.execute('CREATE TABLE t(i INTEGER NOT NULL, a BINARY VARYING NOT NULL, PRIMARY KEY(i))')
+    cur.execute('INSERT INTO t(i, a) VALUES (0, ?)', (b'Hello World',))
     cur.execute('SELECT a FROM t')
     assert cur.fetchone()[0] == b'Hello World'
     assert cur.fetchone() is None
 
+
+def test_insert_double(cur):
+    pytest.skip('Fails with HSQLDB exception')
+    cur.execute('DROP TABLE IF EXISTS abc')
+    cur.execute('CREATE TABLE abc(id INTEGER PRIMARY KEY, a INTEGER)')
+    cur.execute('INSERT INTO abc(id, a) VALUES (1, 2), (?, ?)', (2, 3))
+    cur.execute('SELECT id, a FROM abc ORDER BY id')
+    assert cur.fetchone() == [1, 2]
+    assert cur.fetchone() == [2, 3]
